requirements.txt
streamlit>=1.30.0
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=5.1.0
spacy>=3.7.2
pandas>=2.1.0
rapidfuzz>=3.6.1
reportlab>=4.1.0
python-docx>=1.1.0
PyPDF2>=3.0.1
sentence-transformers>=2.2.2
torch>=2.0.0
